backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pandas==2.2.3
faker==30.8.2
sentence-transformers==3.3.1
python-multipart==0.0.12
requests==2.32.3
python-dotenv==1.0.1
numpy==2.1.3
scikit-learn==1.5.2
huggingface_hub==0.26.2
